Pass template to handler rather than name
By passing the whole template, we can make it easier for developers to
write handlers that might need both the template name relative to the
cwd and to /templates. This is helpful when a user wants to read from a
file (using template.filename) and then write it to its 'proper' spot (using template.name).