Allow passing multiple directories as --static option.
This commit adds support for multiple static directories. They can be passed
as a string of comma-separated names to the CLI and as a list to the
Renderer. This was tested it with a single directory (backwards
compatible) and multiple directories and it works great. If any of the
folders do not exist, it throws the same error as the current version.

Closes #41.