Rename Renderer to Site
Also renames Renderer.run to Site.render. [1]

This change is backwards incompatible, but "Renderer" is semantically
poor and it would be good to change this sooner rather than later. The
fact that rendering is achieved with "Renderer.run()" seems to support
this.

"Site" seems like a better name, since we're rendering a static site.
Jekyll also uses "Site", so there is some precedent.

Closes #37.

.. [1]
    It is worth nothing that Jekyll uses "process", but I think that
    name is semantically empty. Jekyll does use "build" at the command
    line to invoke Site.process, which seems to suggest it was an
    oversight. We use "render" instead of "build" following Jinja2's
    Template.render API.