Make --srcpath work
Presently, using ``--srcpath`` to specify a directory to search for
templates passes in whatever the user types in. Typically, this will be
a relative path (e.g. 'templates'), which will the be coerced to an
absolute by prepending the location of wherever staticjinja is
installed, which is probably not what the user wanted. This is
especially a surprise, because if ``--srcpath`` is not passed in, it
defaults to './templates' and one would expect identical behavior.

This commit fixes changes the behavior of populating of ``--srcpath`` so
that if the path is relative it prepends the current working directory
otherwise it does nothing. If no path is provided, it defaults to the
current working directory + 'templates' as before.

Closes #45.