Improvements to the context workflow
This commit enables users to directly pass dictionaries instead of context
generators in Site (and make_site()) for contexts that don't require any
logic.

It also introduces a "mergecontexts" parameter to Site (and make_site()) to
direct staticjinja to use all matching context generators or only the
first one when rendering templates.

Closes #62.