Add template name to errors when loading template fails
Before, if Jinja2 tried to load a template with a bad encoding (which is
easy to happen if you don't have your static/ignored filters correct and
Jinja2 tries to load a .jpg,.mp3, etc as a template) then we would get
this useless error such as
` 'utf-8' codec can't decode byte 0x80 in position 42: invalid start
byte `
which doesn't say which file or template we were trying to load!

This change throws away some specificity of the error by demoting it
from a UnicodeDecodeError to a UnicodeError, but that was necessary to
be able to add a more detailed message string. I didn't want to just
print the template_name with self.logger.error() because then the
template_name could be printed on a separate line, and I wanted the
template_name to be bundled into the actual error.