Forward staticjinja.make_site() args to Site.make_site() using **kwargs
This will help guarantee that the two functions will always behave identically. For instance, if the default args of one is changed, or an argument is added, then they will both behave the same. I accidentally changed one of the default args and not the other, and this is what caused me to notice this.