Fix and improve Travis CI
The configuration of Travis was brittle. tox only was managing to use python3.4 because it was already installed on Travis instances by luck. This makes it so that Travis explicitly installs a specific python version, and then tox uses that version.

In addition, we were only testing on linux, which was brittle because staticjinja does a large amount of file manipulation. Therefore this adds tests on osx and windows. It's a bit funky because python isn't supported that well on those two OSs, and the range of available of versions is small. osx only has python 3.7 on only osx 10.13 and 10.14. On windows, only python 3.5.4, 3.6.8, and 3.7.4 are available through the chocolatey package manager. Upon adding windows tests it revealed bugs in the tests that ssumed posix style os path separators, so fix those in test_staticjinja.py

Also pass the tox environment through the TOXENV environment variable instead of a command line arg.

Notice that this also drops python2 support. python2 retired jan 1 2020, it's time to stop supporting it!

While were at it, improve the style of tox.ini. In there allow newer versions of flake8.

Finally, in recent versions of pytest the [pytest] section of setup.cfg should be named [tool:pytest] (changed in https://github.com/pytest-dev/pytest/commit/91384193794dcf8ae30778767546a2dad7bad845) so fix that.