(#87, #91) Update custom rendering rules example in docs
See issues #87 and #91 for more context.

I unified the language from `compile` to `render` because that is the term the source code and API uses for this step in the build process. I'm not sure if my definition of the rendering step is very good/complete/clear.

I tweaked @PatrickMassot's example if #87:
1. Use the markdown instead of mistletoe library, I'm more familiar with it and it seems more popular?
2. Do the md->html translation in the context generator because that's simple, the point of this example was to show off the custom render function.
3. Don't use the Path library, instead just use vanilla os.path utilities, as I think they may be mofre familiar to people.
4. Don't use a jinja2 extension, that is sorta out of scope I think? I was just trying to show off custom rendering functions.