(#87) Add markdown example
This is the start of a fix for https://github.com/staticjinja/staticjinja/issues/87, where the current markdown rendering example is totally outdated.

Instead of just fixing the examples embedded in the docs, I wanted to give users a fully featured example that they could run straightaway or copy adn modify. This format of examples folder could also could be used for tests, not sure yet...