Update requirements.txt with piptools
requirements.txt needs to not just hold the direct dependencies of this project, but also the indirect ones. That's nasty to have to actually do.

piptools is a tool which takes your direct dependencies from setup.py, builds a dependency graph using PyPi, and then generates a requirements.txt for you.