Remove broken `filepath` arg from Site.render_templates()
render_templates() passed on the `filepath` arg to render_template() as the second positional argument, meaning it was the interpreted as the `context` arg. This obviously would break render_template() whenever it is used, so I don't think this feature is used.

What's more it doesn't make any sense for it to be used like this, because then every single template would be written to the same output path/stream, overwriting each other.