Enable test code coverage
Add a tox environment `coverage` that runs pytest with the `pytest-cov` plugin that uses the `coverage` package. It makes an HTML view of the results for local debugging, as well as an xml report.

In addition add a `coverage` job to the build workflow on GitHub Actions, that automatically runs code coverage on every PR.