Add examples to tests
While we go to the effort of having exmaples, we could also just use them as test cases. Running each example should transform the input in the way that we expect.

When comparing the `build` and `build-EXPECTED` directories, I tried to print out as much info as possible, so that in the case of a failing test you can pin down the cause easily. This uses the `pytest_check` plugin to `pytest` so you can have multiple assert statements per test, and see the reuslt from all of them.

We have to subclass filecmp.dircmp so that we can do deep file comparisons.

There's some shenanigans of os.chdir()ing to the example working directory to run each example, I hope this is a reasonable way of doing it.