Tweak and update docs/user/quickstart.rst
They stated python versions were out of date. Just delete this specification, it's are bound to always be out of date. You can find that info elsewhere.

Then, make it more explicit that this quickstart is for CLI, and that there is the API for advanced use.

Finally, fix a few style things.