1.0.0 release
It's time to acknowledge that people are relying on this project, so we need to actually bump the major version out of the beta stage. I'm incrementing the major version here because this no longer supports python <3.6, and that could break people's code. There aren't any other breaking API changes though. See the CHANGELOG.rst for detailed info.

This should be auto-released to PyPI based on the workflow I created in the previous commit, but k=who knows I probably pooched it :)