docs: Fix jinja intersphinx mapping
In 0dd4badb94d0446 I changed the intersphinx mapping to https
(lol IDK why that should be important in a docs build...) and that
silently broke the links. Instead of going back to the old http,
might as well upgrade the mapping to Jinja's new website.

While we're at it, add an explicit step in the build CI to test with tox
that the docs build. I thought that seeing the succesful status check
from readthedocs.io was adequate, but looking at the build log for 1.0.2
https://readthedocs.org/projects/staticjinja/builds/12832768/, it shows
the SSL error as a warning, silently failing.