tests: Test docopt part of CLI
Before, when testing the CLI we skipped the docopt parsing
step, passing the parsed arg dictionary into cli.render().
This might have missed some mistakes with how we have docopt
set up, and it's not the use case that users actually see. What users
actually do is pass in a list of args from sys.argv. This refactor
adds a seam for testing in cli.main(), so we can inject custom sys.argv
values, and test the CLI end to end.

While we're here, improve the test cases to use a more obvious CWD.