tests: Cleanup CLI and improve help message
This changes the CLI help message, adding more
details into the optional arguments, and deduplicating
listing them. It also adds info on build vs watch. This relies
on the mechanics of docopt to expand `[options]` into
`--srcpath` etc. This also now includes default values for
srcpath and outpath, both in the help message, and in the args
passed to main().

Adding a default value to `--static` was ugly,
harder to parse in the logic, and technically `--static` is currently
deprecated, even though it still works and I want to fix/improve it.
Just don't do anything about that for now.

Changing these default values made it necessary to change the
logic inside CLI a bit. That's OK though, since it was a bit hard to
follow. This adds a helper method that both srcpath and outpath
can use, which also makes it clear. Now the final result of paths are
always resolved to be absolute, before they weren't necessarily.

Clean up a few things stylewise too.