tests: Add more CLI tests
Add tests for outpath, build vs watch, and nonexistent srcpath

These aspects of the CLI logic was not tested before.
I skipped adding the tests for staticpaths parsing
because at this point that feature is still technically marked
as deprecated, and I hope to overhaul that functionality anyways.

This is in prep to overhaul some of the logic in the CLI, and I want to
ensure that I'm not breaking anything.