tests: Configure flake8 and pytest in setup.cfg
This makes it so tox, CI, and developers just running the commands directly
will all have the same configuration.

IDK what norecursedirs = env was supposed to be doing, but I think it was
garbage. Even at the time it was added, there wasn't an `env` directory
to avoid searching into. Therefore I think it's safe to remove it.