dev: Move checks out of tox into Makefile
Move the docs, flake8, and coverage tasks out of tox.ini
and put them in the Makefile instead. Tox is not necessary
for those tasks, and spinning up the virtual environment is
not needed. When a task is in tox, it's run against a packaged
version of staticjinja. While this is great for running tests, to ensure
that our users are seeing exactly what we do, it's not needed for the
other linters and checks, which be definition should run against the
source code.

Moving them makes it faster (since we don't have to spin up tox) and
just as easy (maybe slightly easier?) to run these tasks.

It also makes the docs build to a more sane place outside of the .tox
directory, and allows us to avoid the opaque {envtmpdir} location of tox.

Making `make test` be the dev testing entry point also allows us to
improve how we were dealing with tox's --skip_missing_interpreters
argument. Things are maybe safer now, with tox failing by default
unless all the proper interpreters are available.

This also adds some helper make recipes to view the coverage
or docs after building them, as well as uploading coverage reports.
Make's reciper dependencies are really useful here.

Now CI calls into make directly instead of calling tox, unifying what
devs and CI need to do. The lint job was renamed to flake8 in CI.

Now `make test` also checks that the package builds.

Now building docs also verifies that all external links are still valid.
This caught some errors so they get fixed here too.