tests: Fix and improve flake8
The configuration of flake8 in setup.cfg to look for certain
`filename`s was wrong, and was making flake8 not actually find
any files to test. One option would be to fix the glob patterns to use
`./tests/*.py` and similar, but that would be fragile to changed or new
filenames (for example, setup.py doesn't even exist at this point but
we stall had it listed...). Therefore, just get rid of the filename config
totally, and flake8 will default to `*.py`, which is reasonable. All python
files should probably be linted. With this change it even caught
some errors that were easy to fix.