core: Simplify ensuring that out directories exists
See https://github.com/staticjinja/staticjinja/issues/64#issuecomment-778715178
The two old checks were redundant and stepped on each others toes.
This consolidates them into one check, and it does it more simply. Now
the callers of the check give the output location, instead of the input.
This makes the logic much simpler, especially since the calculation of
the output location was done in the callers anyway.

This *does* change the behavior, reverting the change made in
https://github.com/staticjinja/staticjinja/pull/71. I don't think that change
was a good direction, because a custom render rule might not
necessarilly map an input file location straight to the same location
under the build directory. The responsibility belongs to the custom
render rule to both do the rendering and the writing to disk,
however it sees fit.

The test_render_templates() test covers the case where a template
should end up in a subdirectory of the out directory. Where this is
used in copy_static() is not tested, but that feature is currently
deprecated and its not even tested at all. OK, if I'm being honest
this change could probably use more testing, but I'm fairly confident
that this doesn't break anything.