Reloader: Improve should_handle()
Use pathlib for easier path manipulation, and to make the
filename argument accept more types.

In addition, the check if the file was in searchpath is unneeded,
so remove it. By definition when we call easywatch.watch() on
searchpath, every callback must be for a file under search path.
In addition, in event_handler() we ASSUME that file is under
search path, since we call os.path.relpath(src_path, searchpath).
That call would explode if src_path wasn't under searchpath.