Site: Make many functions accept PathLike args
Many of the functions in Site would make sense to accept
PathLike objects as arguments. In particular, is_partial()
(and similar) are used for testing what to do with files, and
these functions are called from both get_dependencies()
(which is assuming file paths) and from template_names()
(which is calling the filter functions with jinja2 posix '/'
separators). Either way, the argument represents a path.

This does directly revert many of the changes in
https://github.com/staticjinja/staticjinja/pull/104, but I
am thinking now that that was going in the wrong
direction. This should free up client code to call these
filter and helpt functions with whatever they want
(and I don't think it's allowing them to use arguments
that don't make sense?)