Reloader: Improve Site.get_dependencies()
First rename it to get_dependents(), since that's what we're
actually doing. Add in a shim redirect so we don't break anyone
but deprecate the old name.

Second, fix it so that it always returns a list of filenames.
Before the return type was either a list of templates or
list of filenames, and Reloader used them inconsistently.
With this change Reloader's event_handler() is I think more
clear too, since it always can expect a list of filenames from
get_dependents.

Third, improve get_dependents()'s docstring to explain how
dependents are calculated.

I had to tweak a test because Reloader is now calling Site.copy_static()
with pathlib.Path's instead of just plain strings.