tests: Change CWD to tmp_path before each test
The functionality of staticjinja, especially how it interprets
relative paths, is heavily tied to what is considered the
current working directory. Therefore, we now do
`os.chdir(tmp_path)` before each test (and revert that during
teardown) so that it's like the user is running staticjinja
from the root of their project.