refactor: Make inspect.stack() logic clearer
There was a bunch of opaque indexing into objects,
now the logic is split into multiple statements
with some intermediate named variables.

As a side note, I'm not sure I like this logic. I think
it may be better to always use `os.getcwd()` as the project_path,
instead of using the directory of the calling module. I'm not
positive, but I believe that os.getcwd() is generally considered
a more standard "root". Will consider more.