docs: Fix docstring for Site.render_template
The `filepath` arg only can be a pathlike object, since we
call `_ensure_dir()` on it. The `template.stream(...)` call would
have been fine, it's just our code that restricts the type.
I figure it's not worth it to also support file-like objects,
if people want that, please open an issue/PR.