tests: Add tests for followlinks
This argument to make_site() was never tested. I doubt it's
used that much so it probably isn't tested in the wild that much.
I think in an ideal world I'd like to outsource this responsibility
somehow, making it so that if a user wants this custom
jinja.FileSystemLoader functionality then they need to inject
their own FileSystemLoader.

But, for now I'll just add these tests, which will serve mostly
as a checkpoint, so if in the future this functionality changes
accidentally, we will catch it. I decided to add this when I
started working on
https://github.com/staticjinja/staticjinja/issues/149
and I started really messing with how the searchpath etc are resolved.
I didn't want to accidentally regress anything.