Make examples use a "before" and "after" structure
It was a bit confusing how to interpret the examples before.
Plus, it limited us to always use `build/` as the output folder,
which limited us in testing for
https://github.com/staticjinja/staticjinja/issues/160
This new format is more flexible, and I think more intuitive.

While I'm in the test code, switch us over to using
tmp_path instead of the deprecated tmpdir
and use more of pathlib